requirements.txt
beautifulsoup4==4.11.2
discord.py==2.2.2
openai==0.27.4
PyNaCl==1.5.0
requests==2.28.2
urllib3==1.26.15
yt-dlp==2023.3.4
opuslib==3.0.1
ffmpeg==1.4